@DeleteMapping,@PutMapping

+++ src/main/java/com/example/pizzaonline/repositories/PizzaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PizzaRepository extends JpaRepository<Pizza, Integer> {
+public interface PizzaRepository extends JpaRepository<Pizza, Long> {
     Pizza findByName(String name);
 }
